Split LGraphNode out to TS file
